imagine/pyproject.toml
[project]
name = "imagine"
version = "0.1.0"
description = ""
authors = [
    {name = "hatvalues",email = "julian@hatvalues.org"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "pinterest-api (>=0.0.8,<0.0.9)",
    "requests (>=2.32.3,<3.0.0)",
    "scrapy (>=2.12.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
]

[tool.poetry]
packages = [{include = "imagine", from = "src"}]
package-mode = false


[tool.poetry.group.test.dependencies]
pytest-cov = "^6.1.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"
pylint = "^3.3.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
